Lots of fixes and improvements

+++ src/Controllers/CategoryController.php
@@ -17,8 +17,10 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', Category::class);
+
         $categories = Category::all();
-        return view('laralum_forum::categories.index', ['categories' => $categories]);
+        return view('laralum_forum::laralum.categories.index', ['categories' => $categories]);
     }
 
     /**
@@ -28,7 +30,9 @@ class CategoryController extends Controller
      */
     public function create()
     {
-        return view('laralum_forum::categories.create');
+        $this->authorize('create', Category::class);
+
+        return view('laralum_forum::laralum.categories.create');
     }
 
     /**
@@ -39,12 +43,13 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Category::class);
+
         $this->validate($request, [
-            'title' => 'required|min:5|max:50',
-            'description' => 'required|max:100',
+            'name' => 'required|max:255',
         ]);
         Category::create($request->all());
-        return redirect()->route('laralum::forum.categories.index')->with('success', __('laralum_forum::general.category_created'));
+        return redirect()->route('laralum::forum.categories.index')->with('success', __('laralum_forum::category_added'));
     }
 
     /**
@@ -55,7 +60,9 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
-        return view('laralum_forum::categories.show', ['category' => $category]);
+        $this->authorize('view', Category::class);
+
+        return view('laralum_forum::laralum.categories.show', ['category' => $category]);
     }
 
     /**
@@ -66,7 +73,9 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
-        return view('laralum_forum::categories.edit', ['category' => $category]);
+        $this->authorize('update', $category);
+
+        return view('laralum_forum::laralum.categories.edit', ['category' => $category]);
     }
 
     /**
@@ -78,12 +87,13 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        $this->authorize('update', $category);
+
         $this->validate($request, [
-            'title' => 'required|min:5|max:50',
-            'description' => 'required|max:100',
+            'name' => 'required|max:255',
         ]);
         $category->update($request->all());
-        return redirect()->route('laralum::forum.categories.index')->with('success', __('laralum_forum::general.category_updated',['id' => $category->id]));
+        return redirect()->route('laralum::forum.categories.index')->with('success', __('laralum_forum::category_updated',['id' => $category->id]));
     }
 
     /**
@@ -94,10 +104,11 @@ class CategoryController extends Controller
      */
     public function confirmDestroy(Category $category)
     {
+        $this->authorize('delete', $category);
 
         return view('laralum::pages.confirmation', [
             'method' => 'DELETE',
-            'message' => __('laralum_forum::general.sure_del_category', ['category' => $category->title]),
+            'message' => __('laralum_forum::general.sure_del_category', ['category' => $category->name]),
             'action' => route('laralum::forum.categories.destroy', ['category' => $category->id]),
         ]);
     }
@@ -110,6 +121,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
+
         $category->deleteComments();
         $category->deleteThreads();
         $category->delete();
